Refactor to simple getAppConfig method that can be used on any chain
