+++ apps/hyperdrive-trading/src/appconfig/chains/supportedChains.ts
